fixe

+++ src/find_intercepts_steps.c
@@ -19,7 +19,7 @@ void	find_horz_intercepts(t_ray_wall_hit_horz *horz_ray_data,
 	if (ray_facing_data->is_ray_facing_down)
 		horz_ray_data->y_intercept += CUBE_SIZE;
 	horz_ray_data->x_intercept = player->x + (horz_ray_data->y_intercept
-			- player->y) / tan(ray_angle);
+			- player->y) / tanf(ray_angle);
 	horz_ray_data->next_horz_touch_wall_x = horz_ray_data->x_intercept;
 	horz_ray_data->next_horz_touch_wall_y = horz_ray_data->y_intercept;
 }
@@ -30,7 +30,7 @@ void	find_horz_steps(t_ray_wall_hit_horz *horz_ray_data, float ray_angle,
 	horz_ray_data->y_step = CUBE_SIZE;
 	if (ray_facing_data->is_ray_facing_up)
 		horz_ray_data->y_step *= -1;
-	horz_ray_data->x_step = CUBE_SIZE / tan(ray_angle);
+	horz_ray_data->x_step = CUBE_SIZE / tanf(ray_angle);
 	if (ray_facing_data->is_ray_facing_left && horz_ray_data->x_step > 0)
 		horz_ray_data->x_step *= -1;
 	if (ray_facing_data->is_ray_facing_right && horz_ray_data->x_step < 0)
@@ -44,7 +44,7 @@ void	find_vert_intercepts(t_ray_wall_hit_vert *vert_ray_data,
 	if (ray_facing_data->is_ray_facing_right)
 		vert_ray_data->x_intercept += CUBE_SIZE;
 	vert_ray_data->y_intercept = player->y + (vert_ray_data->x_intercept
-			- player->x) * tan(ray_angle);
+			- player->x) * tanf(ray_angle);
 	vert_ray_data->next_vert_touch_wall_x = vert_ray_data->x_intercept;
 	vert_ray_data->next_vert_touch_wall_y = vert_ray_data->y_intercept;
 }
@@ -55,7 +55,7 @@ void	find_vert_steps(t_ray_wall_hit_vert *vert_ray_data, float ray_angle,
 	vert_ray_data->x_step = CUBE_SIZE;
 	if (ray_facing_data->is_ray_facing_left)
 		vert_ray_data->x_step *= -1;
-	vert_ray_data->y_step = CUBE_SIZE * tan(ray_angle);
+	vert_ray_data->y_step = CUBE_SIZE * tanf(ray_angle);
 	if (ray_facing_data->is_ray_facing_up && vert_ray_data->y_step > 0)
 		vert_ray_data->y_step *= -1;
 	if (ray_facing_data->is_ray_facing_down && vert_ray_data->y_step < 0)
